feat: api 명세 변경에 따른 mockData 추가 및 타입명 변경
